requirements.txt
# Django e extensões principais
Django==4.2.7
django-environ==0.11.2
psycopg2-binary==2.9.8
django-extensions==3.2.3

# Interface e formulários
django-crispy-forms==2.0
django-widget-tweaks==1.5.0
crispy-bootstrap5==0.7

# Geolocalização e mapas
django-leaflet==0.29.0
geopy==2.4.0

# Dados e APIs
requests==2.31.0
pandas==2.1.3

# Agendamento de tarefas
APScheduler==3.10.4

# Cache e performance
redis==5.0.1
django-redis==5.4.0

# Desenvolvimento e debug
django-debug-toolbar==4.2.0

# Segurança
django-cors-headers==4.3.1

# Testes
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0

# Utilitários
python-decouple==3.8
Pillow==10.1.0
whitenoise==6.6.0
gunicorn==21.2.0 